% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Title}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{lfun is sample distribution}

\item{x}{vector of the sample}

\item{param}{name of the parameter}

\item{...}{}
}
\value{
it returns graph for the given distribution
}
\description{
Title
}
