% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ticketApprox}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seat}

\item{gamma}{probability of overbooking completely}

\item{p}{probability of show up}
}
\value{
returns 2 way estimated values for number of tickets to sell and  graph of discrete and continuous
}
\description{
ticketApprox
}
\examples{
\dontrun{ntickets(200,0.02,0.95)}

}
