% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.R
\docType{data}
\name{spruce}
\alias{spruce}
\title{Height and Diameter of Spruce frees}
\format{
A data frame with 144 rows and variables
\describe{
  \item{Height}{Height in meters}
  \item{BHDiameter}{Breast Height Diameter in cm}
  ...
}
}
\source{
\url {https://mnspruce.ornl.gov/datasets}
}
\usage{
spruce
}
\description{
144 trees
}
\keyword{datasets}
